Add demos with SVM/NN analysis
This commit adds several .Rmd files in the demo directory that were used to train machine learning models on corruption. The results of these calculations are in the data-raw directory; there are subdirectories for search-grids and results-scores (results-models is also a directory but is excluded from git, as the file sizes are too large at this time). results-scores has an images subdirectory that help visualize the hyperparameter tuning for the SVM/NN analysis. This commit also changed the markdown format of the vignettes to github_document.